+++ fitsio/fitsio_pywrap.c
@@ -1471,6 +1471,7 @@ static PyObject *PyFITSObject_create_image_hdu(struct PyFITSObject *self,
     int comptype = 0;
     int orig_dither_seed;
     int got_dither_seed = 0;
+    int any_nan = 0;
 
     if (self->fits == NULL) {
         PyErr_SetString(PyExc_ValueError, "fits file is NULL");
@@ -1488,20 +1489,15 @@ static PyObject *PyFITSObject_create_image_hdu(struct PyFITSObject *self,
     }
 
     static char *kwlist[] = {
-        "array",       "nkeys",       "dims",
-
-        "comptype",    "tile_dims",   "qlevel",       "qmethod",
-        "dither_seed", "hcomp_scale", "hcomp_smooth",
-
-        "extname",     "extver",      NULL,
+        "array",   "nkeys",   "dims",        "comptype",    "tile_dims",
+        "qlevel",  "qmethod", "dither_seed", "hcomp_scale", "hcomp_smooth",
+        "extname", "extver",  "any_nan",     NULL,
     };
     if (!PyArg_ParseTupleAndKeywords(
-            args, kwds, "Oi|OOOOOOOOsi", kwlist, &array_obj, &nkeys, &dims_obj,
-
+            args, kwds, "Oi|OOOOOOOOsii", kwlist, &array_obj, &nkeys, &dims_obj,
             &comptype_obj, &tile_dims_obj, &qlevel_obj, &qmethod_obj,
-            &dither_seed_obj, &hcomp_scale_obj, &hcomp_smooth_obj,
-
-            &extname, &extver)) {
+            &dither_seed_obj, &hcomp_scale_obj, &hcomp_smooth_obj, &extname,
+            &extver, &any_nan)) {
         py_status = 1;
         goto create_image_hdu_cleanup;
     }
@@ -1661,10 +1657,30 @@ static PyObject *PyFITSObject_create_image_hdu(struct PyFITSObject *self,
         void *data = NULL;
         nelements = PyArray_SIZE(array);
         data = PyArray_DATA(array);
-        if (fits_write_img(self->fits, datatype, firstpixel, nelements, data,
-                           &status)) {
-            set_ioerr_string_from_status(status, self);
-            goto create_image_hdu_cleanup;
+        if (any_nan) {
+            float fnullval = INFINITY;
+            double dnullval = INFINITY;
+            void *nullval_ptr;
+
+            if (datatype == TFLOAT) {
+                nullval_ptr = (void *)(&fnullval);
+            } else if (datatype == TDOUBLE) {
+                nullval_ptr = (void *)(&dnullval);
+            } else {
+                nullval_ptr = NULL;
+            }
+
+            if (fits_write_imgnull(self->fits, datatype, firstpixel, nelements,
+                                   data, nullval_ptr, &status)) {
+                set_ioerr_string_from_status(status, self);
+                goto create_image_hdu_cleanup;
+            }
+        } else {
+            if (fits_write_img(self->fits, datatype, firstpixel, nelements,
+                               data, &status)) {
+                set_ioerr_string_from_status(status, self);
+                goto create_image_hdu_cleanup;
+            }
         }
     }
 
@@ -1760,6 +1776,8 @@ static PyObject *PyFITSObject_write_image(struct PyFITSObject *self,
     LONGLONG nelements = 1;
     PY_LONG_LONG firstpixel_py = 0;
     LONGLONG firstpixel = 0;
+    PY_LONG_LONG any_nan_py = 0;
+    LONGLONG any_nan = 0;
     int image_datatype = 0; // fits type for image, AKA bitpix
     int datatype = 0;       // type for the data we entered
 
@@ -1774,8 +1792,8 @@ static PyObject *PyFITSObject_write_image(struct PyFITSObject *self,
         return NULL;
     }
 
-    if (!PyArg_ParseTuple(args, (char *)"iOL", &hdunum, &array_obj,
-                          &firstpixel_py)) {
+    if (!PyArg_ParseTuple(args, (char *)"iOLL", &hdunum, &array_obj,
+                          &firstpixel_py, &any_nan_py)) {
         return NULL;
     }
     array = (PyArrayObject *)array_obj;
@@ -1798,12 +1816,33 @@ static PyObject *PyFITSObject_write_image(struct PyFITSObject *self,
     data = PyArray_DATA(array);
     nelements = PyArray_SIZE(array);
     firstpixel = (LONGLONG)firstpixel_py;
-    if (fits_write_img(self->fits, datatype, firstpixel, nelements, data,
-                       &status)) {
-        set_ioerr_string_from_status(status, self);
-        return NULL;
-    }
+    any_nan = (LONGLONG)any_nan_py;
 
+    if (any_nan_py) {
+        float fnullval = INFINITY;
+        double dnullval = INFINITY;
+        void *nullval_ptr;
+
+        if (datatype == TFLOAT) {
+            nullval_ptr = (void *)(&fnullval);
+        } else if (datatype == TDOUBLE) {
+            nullval_ptr = (void *)(&dnullval);
+        } else {
+            nullval_ptr = NULL;
+        }
+
+        if (fits_write_imgnull(self->fits, datatype, firstpixel, nelements,
+                               data, nullval_ptr, &status)) {
+            set_ioerr_string_from_status(status, self);
+            return NULL;
+        }
+    } else {
+        if (fits_write_img(self->fits, datatype, firstpixel, nelements, data,
+                           &status)) {
+            set_ioerr_string_from_status(status, self);
+            return NULL;
+        }
+    }
     // this is a full file close and reopen
     if (fits_flush_file(self->fits, &status)) {
         set_ioerr_string_from_status(status, self);
@@ -4218,6 +4257,18 @@ recread_asrec_cleanup:
     Py_RETURN_NONE;
 }
 
+static int fits_is_compressed_with_nulls(fitsfile *fits) {
+    int cstatus = 0, clstatus = 0, hstatus = 0;
+    int colnum, zblank;
+    if (fits_is_compressed_image(fits, &cstatus) &&
+        (ffgcno(fits, CASEINSEN, "ZBLANK", &colnum, &clstatus) <= 0 ||
+         ffgky(fits, TINT, "ZBLANK", &zblank, NULL, &hstatus) <= 0)) {
+        return 1;
+    } else {
+        return 0;
+    }
+}
+
 // read an n-dimensional "image" into the input array.  Only minimal checking
 // of the input array is done.
 // Note numpy allows a maximum of 32 dimensions
@@ -4283,11 +4334,27 @@ static PyObject *PyFITSObject_read_image(struct PyFITSObject *self,
     npy_dtype = PyArray_TYPE(array);
     npy_to_fits_image_types(npy_dtype, &dummy, &fits_read_dtype);
 
+    float fnullval = NAN;
+    double dnullval = NAN;
+    void *nullval_ptr = NULL;
+
+    // we only set null checking for compressed images of
+    // floating point data
+    // nans works fine for non-compressed images and we do
+    // not consider int data
+    if (fits_is_compressed_with_nulls(self->fits)) {
+        if (fits_read_dtype == TFLOAT) {
+            nullval_ptr = (void *)(&fnullval);
+        } else if (fits_read_dtype == TDOUBLE) {
+            nullval_ptr = (void *)(&dnullval);
+        }
+    }
+
     for (i = 0; i < naxis; i++) {
         firstpixels[i] = 1;
     }
-    if (fits_read_pixll(self->fits, fits_read_dtype, firstpixels, size, 0, data,
-                        &anynul, &status)) {
+    if (fits_read_pixll(self->fits, fits_read_dtype, firstpixels, size,
+                        nullval_ptr, data, &anynul, &status)) {
         set_ioerr_string_from_status(status, self);
         return NULL;
     }
@@ -4438,8 +4505,24 @@ static PyObject *PyFITSObject_read_image_slice(struct PyFITSObject *self,
     npy_dtype = PyArray_TYPE((PyArrayObject *)array);
     npy_to_fits_image_types(npy_dtype, &dummy, &fits_read_dtype);
 
-    if (fits_read_subset(self->fits, fits_read_dtype, fpix, lpix, step, 0, data,
-                         &anynul, &status)) {
+    float fnullval = NAN;
+    double dnullval = NAN;
+    void *nullval_ptr = NULL;
+
+    // we only set null checking for compressed images of
+    // floating point data
+    // nans works fine for non-compressed images and we do
+    // not consider int data
+    if (fits_is_compressed_with_nulls(self->fits)) {
+        if (fits_read_dtype == TFLOAT) {
+            nullval_ptr = (void *)(&fnullval);
+        } else if (fits_read_dtype == TDOUBLE) {
+            nullval_ptr = (void *)(&dnullval);
+        }
+    }
+
+    if (fits_read_subset(self->fits, fits_read_dtype, fpix, lpix, step,
+                         nullval_ptr, data, &anynul, &status)) {
         set_ioerr_string_from_status(status, self);
         goto read_image_slice_cleanup;
     }
@@ -4883,6 +4966,10 @@ static PyObject *PyFITS_cfitsio_has_curl_support(void) {
 #endif
 }
 
+static PyObject *PyFITS_cfitsio_null_value_for_nan(void) {
+    return PyFloat_FromDouble((double)INFINITY);
+}
+
 /*
 
 'C',              'L',     'I',     'F'             'X'
@@ -5186,6 +5273,10 @@ static PyMethodDef fitstype_methods[] = {
      METH_NOARGS,
      "cfitsio_has_curl_support\n\nReturn True if cfitsio has support for "
      "curl."},
+    {"cfitsio_null_value_for_nan",
+     (PyCFunction)PyFITS_cfitsio_null_value_for_nan, METH_NOARGS,
+     "cfitsio_null_value_for_nan\n\nReturn our default null value for "
+     "floats, which is INFINITY and/or np.inf"},
     {"parse_card", (PyCFunction)PyFITS_parse_card, METH_VARARGS,
      "parse_card\n\nparse the card to get the key name, value (as a string), "
      "data type and comment."},